add user register, login, refresh, logout

+++ src/services/contacts.js
@@ -9,11 +9,12 @@ export const getContacts = async ({
   sortOrder = SORT_ORDER.ASC,
   sortBy = '_id',
   filter = {},
+  userId,
 }) => {
   const limit = perPage;
   const skip = (page - 1) * perPage;
 
-  const contactsQuery = Contacts.find();
+  const contactsQuery = Contacts.find({ userId });
   if (filter.contactType) {
     contactsQuery.where('contactType').equals(filter.contactType);
   }
@@ -23,7 +24,7 @@ export const getContacts = async ({
   }
 
   const [contactsCount, contacts] = await Promise.all([
-    Contacts.find().merge(contactsQuery).countDocuments(),
+    Contacts.find({ userId }).merge(contactsQuery).countDocuments(),
     contactsQuery
       .skip(skip)
       .limit(limit)
@@ -39,17 +40,22 @@ export const getContacts = async ({
     ...paginationData,
   };
 };
-export const getContactById = async (contactId) => {
-  const contact = await Contacts.findById(contactId);
+export const getContactById = async (contactId, userId) => {
+  const contact = await Contacts.findOne({ _id: contactId, userId });
   return contact;
 };
 export const postContact = async (payload) => {
   const contact = await Contacts.create(payload);
   return contact;
 };
-export const updateContact = async (contactId, payload, options = {}) => {
+export const updateContact = async (
+  contactId,
+  payload,
+  userId,
+  options = {},
+) => {
   const rawResult = await Contacts.findOneAndUpdate(
-    { _id: contactId },
+    { _id: contactId, userId },
     payload,
     {
       new: true,
@@ -65,7 +71,7 @@ export const updateContact = async (contactId, payload, options = {}) => {
     isNew: Boolean(rawResult?.lastErrorObject?.upserted),
   };
 };
-export const deleteContact = async (contactId) => {
-  const contact = await Contacts.findOneAndDelete({ _id: contactId });
+export const deleteContact = async (contactId, userId) => {
+  const contact = await Contacts.findOneAndDelete({ _id: contactId, userId });
   return contact;
 };
